[Symfony] Fix: Set span status according to otel convention (#295)
* Set span status according to otel convention

The span status was always set to error for responses with status code 400 and up. 
According to the otel semantic conventions, for 4xx status, the span status MUST be left unset for spans with SpanKind.SERVER
This commit adds the logic to implement this behaviour.

Ref: https://opentelemetry.io/docs/specs/semconv/http/http-spans/#status

* Do not set error status on internal spans with status 4xx

Internal spans could have an error status, while the accompanying server span didn't have an error status.
This changes this behaviour to ONLY set error status on requests with status code >= 500 regardless of the span kind.